# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: attendance
    description: Employee attendance operations
  - name: admin
    description: Administrative attendance operations
  - name: auth
    description: Local account and session operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /attendance/today:
    get:
      operationId: getTodayAttendance
      tags: [attendance]
      summary: Get the current employee attendance state
      responses:
        "200":
          description: Current attendance state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceStatus"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Start a local employee session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          description: Invalid credentials
        "429":
          description: Too many attempts
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: End the current local session
      responses:
        "204":
          description: Session cleared
  /auth/me:
    get:
      operationId: getCurrentSession
      tags: [auth]
      summary: Get current local session
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
  /attendance/scan:
    post:
      operationId: scanAttendanceQr
      tags: [attendance]
      summary: Consume a rotating QR token and record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceScanInput"
      responses:
        "200":
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceEvent"
        "400":
          description: Invalid, expired, or already-used token
        "429":
          description: Too many scan attempts
  /admin/qr:
    get:
      operationId: getQrStatus
      tags: [admin]
      summary: Get the currently active QR token status
      responses:
        "200":
          description: QR status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrStatus"
  /admin/qr/rotate:
    post:
      operationId: rotateQrToken
      tags: [admin]
      summary: Rotate the active QR token
      responses:
        "200":
          description: New QR status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrStatus"
  /admin/attendance:
    get:
      operationId: listAttendanceEvents
      tags: [admin]
      summary: List recent attendance events
      parameters:
        - name: date
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Attendance events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceEvent"
  /admin/summary:
    get:
      operationId: getAttendanceSummary
      tags: [admin]
      summary: Get attendance dashboard summary
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"
  /admin/employees:
    get:
      operationId: listEmployees
      tags: [admin]
      summary: List local employee accounts
      responses:
        "200":
          description: Employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [admin]
      summary: Create a local employee account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeCreateInput"
      responses:
        "201":
          description: Created employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /admin/employees/{id}:
    put:
      operationId: updateEmployee
      tags: [admin]
      summary: Update an employee account
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdateInput"
      responses:
        "200":
          description: Updated employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AttendanceScanInput:
      type: object
      properties:
        token:
          type: string
          minLength: 16
          maxLength: 512
        selfie:
          type: string
          minLength: 100
          description: Front-camera image as a JPG, PNG, or WebP data URL.
      required: [token, selfie]
    AttendanceEvent:
      type: object
      properties:
        id: { type: string }
        employeeId: { type: string }
        employeeName: { type: string }
        type:
          type: string
          enum: [check_in, check_out]
        timestamp: { type: string, format: date-time }
        location:
          type: ["string", "null"]
        deviceLabel:
          type: ["string", "null"]
        selfieUrl:
          type: ["string", "null"]
        loginAt:
          type: ["string", "null"]
          format: date-time
      required: [id, employeeId, employeeName, type, timestamp, location, deviceLabel, selfieUrl, loginAt]
    AttendanceStatus:
      type: object
      properties:
        employeeId: { type: string }
        employeeName: { type: string }
        state:
          type: string
          enum: [out, checked_in, checked_out]
        checkIn:
          type: ["string", "null"]
          format: date-time
        checkOut:
          type: ["string", "null"]
          format: date-time
        workedMinutes: { type: number }
      required: [employeeId, employeeName, state, checkIn, checkOut, workedMinutes]
    QrStatus:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        rotatedAt:
          type: string
          format: date-time
        remainingSeconds:
          type: number
      required: [token, expiresAt, rotatedAt, remainingSeconds]
    AttendanceSummary:
      type: object
      properties:
        present: { type: number }
        expected: { type: number }
        late: { type: number }
        checkedOut: { type: number }
        lastEvent:
          oneOf:
            - $ref: "#/components/schemas/AttendanceEvent"
            - type: "null"
      required: [present, expected, late, checkedOut, lastEvent]
    LoginInput:
      type: object
      properties:
        username: { type: string, minLength: 1, maxLength: 80 }
        password: { type: string, minLength: 1, maxLength: 256 }
      required: [username, password]
    SessionEmployee:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        role: { type: string, enum: [admin, employee] }
        profilePhotoUrl: { type: ["string", "null"] }
      required: [id, username, displayName, role, profilePhotoUrl]
    SessionResponse:
      type: object
      properties:
        employee: { $ref: "#/components/schemas/SessionEmployee" }
        loginAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
      required: [employee, expiresAt]
    Employee:
      type: object
      properties:
        id: { type: string }
        username: { type: string }
        displayName: { type: string }
        documentNumber: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        jobTitle: { type: ["string", "null"] }
        active: { type: boolean }
        role: { type: string, enum: [admin, employee] }
        profilePhotoUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, username, displayName, documentNumber, email, phone, jobTitle, active, role, profilePhotoUrl, createdAt]
    EmployeeCreateInput:
      type: object
      properties:
        username: { type: string, minLength: 1, maxLength: 80 }
        password: { type: string, minLength: 12, maxLength: 256 }
        displayName: { type: string, minLength: 1, maxLength: 160 }
        documentNumber: { type: string, maxLength: 80 }
        email: { type: string, maxLength: 160 }
        phone: { type: string, maxLength: 80 }
        jobTitle: { type: string, maxLength: 120 }
        profilePhoto: { type: string, maxLength: 3000000 }
      required: [username, password, displayName]
    EmployeeUpdateInput:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 160 }
        documentNumber: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        jobTitle: { type: ["string", "null"] }
        password: { type: string, minLength: 12, maxLength: 256 }
        profilePhoto: { type: string, maxLength: 3000000 }
        active: { type: boolean }