# lib/api-zod/src/generated/types/attendanceEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceEventType } from './attendanceEventType';

export interface AttendanceEvent {
  id: string;
  employeeId: string;
  employeeName: string;
  type: AttendanceEventType;
  timestamp: Date;
  /** @nullable */
  location: string | null;
  /** @nullable */
  deviceLabel: string | null;
  /** @nullable */
  selfieUrl: string | null;
  /** @nullable */
  loginAt: Date | null;
}