# lib/api-zod/src/generated/types/attendanceScanInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttendanceScanInput {
  /**
     * @minLength 16
     * @maxLength 512
     */
  token: string;
  /**
     * Front-camera image as a JPG, PNG, or WebP data URL.
     * @minLength 100
     */
  selfie: string;
}