# lib/api-zod/src/generated/types/sessionEmployee.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionEmployeeRole } from './sessionEmployeeRole';

export interface SessionEmployee {
  id: string;
  username: string;
  displayName: string;
  role: SessionEmployeeRole;
  /** @nullable */
  profilePhotoUrl: string | null;
}